pyproject.toml
[tool.poetry]
name = "zyuzyubliki-test-task"
version = "0.1.0"
description = ""
authors = ["Artemii Berezin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.5"
SQLAlchemy = "^2.0.20"
aiosqlite = "^0.19.0"
alembic = "^1.12.0"
pandas = "^2.1.0"
init = "^0.1.0"
pip = "^23.2.1"
python-dotenv = "^1.0.0"
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
greenlet = "^2.0.2"
openpyxl = "^3.1.2"
prettytable = "^3.8.0"
fake-useragent = "^1.2.1"
aiohttp = "^3.8.5"
lxml = "^4.9.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
